добавил ShowRate Ctr в модель ответа GetStatistic,переименовал без Total,

+++ XInstallBotProfile/Models/UserStatistic.cs
@@ -13,10 +13,10 @@ namespace XInstallBotProfile.Models
         public long ImpsCount { get; set; }
         public decimal ShowRate { get; set; }
         public long ClicksCount { get; set; }
-        public long Ctr { get; set; }
+        public decimal Ctr { get; set; }
         public long StartsCount { get; set; }
         public long CompletesCount { get; set; }
-        public long Vtr { get; set; }
+        public decimal Vtr { get; set; }
 
         public bool IsDsp { get; set; }
         public bool IsDspInApp { get; set; }
